Making the example run slightly more streamlined.
We only do the mergePipe once, as only a maximum of two workers are
ever used. Indeed, the reading stage is generally faster than the
merging stage and so after a while, we are once more reduced to a
sequential operation. This will not change no matter how many cores
you throw at it.